refactor: Move API exceptions to api/exceptions.py
No error data classes exist in this layer so errors.py is removed entirely.
