+++ sharedkernel/api/__init__.py
@@ -1,0 +1,5 @@
+from . import exceptions
+
+__all__ = [
+    "exceptions",
+]
